requirements.txt
loguru
mss
numpy
opencv-python
pillow
pyautogui
PyGetWindow
PyYAML
keyboard
paddleocr
